Rename color_block to block
