+++ src/block.cpp
@@ -1,4 +1,4 @@
-#include "color_block.hpp"
+#include "block.hpp"
 
 BlockBase::BlockBase() {}
 BlockBase::~BlockBase() {}
