+++ src/main.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
+#include "block.hpp"
 #include "codel.hpp"
 #include "codel_table.hpp"
-#include "color_block.hpp"
 #include "interpreter.hpp"
 #include "png.hpp"
 #include "visualize.hpp"
